making authorized requests

+++ frontend/src/components/WorkoutsDetails.js
@@ -1,13 +1,18 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 import formatDistantToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutsDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) return;
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: { 'Authorization': `Bearer ${user.Authorization}` },
     });
     const data = await response.json();
